requirements.txt
faiss-cpu
numpy
sentence-transformers
transformers
fastapi
uvicorn
pydantic
python-dotenv
huggingface_hub